matiere

+++ app_gestion_note/src/main/java/com/example/demo/entities/Matiere.java
@@ -24,7 +24,7 @@ public class Matiere {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long idMatiere;
 	private String nomMatiere;
-	private Long CoifMatiere;
+	private int CoifMatiere;
 	@ManyToOne
 	@JsonIgnore
 	private Module M;
